# Optimize filebeat state ID generation