@ruflin 2018-02-11
Thanks for the benchmarks. This is really useful.

Should we add a changelog entry for this as this is quite an enhancement? 

CI failures were not related.