@urso 2018-02-12
No changelog, as It's not really a much 'visible' change.